backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.2.0
gunicorn==21.2.0
drf-yasg==1.21.7
